test: migrate remaining webview specs to main
